Correct RS1XX Presentation Issues
This corrects temperature ranges not being
displayed correctly for all sensor types. This
is due to the omission of the Operating
Conditions field in all sensor .yaml files.

Also corrects \# appearing in sensor
descriptions due to inclusion of # character
but lack of inverted commas.

Requested by: <greg.leach@lairdconnect.com>