Fix decoder for LGT-92 fw version 1.5/1.6 devices
Especially, it fixes the issue where longitude/latitude are set to zero
if the MSB is zero.

It also fixes the decoding of the Move-Direction (MD) field, where e.g.
'Move' was erroneously decoded as 'Disable', before.

While at it I also added some warnings for some special conditions.

It's a complete rewrite of the previous decoder in order to make it
shorter, easier reviewable and arguably simpler to review.

fixes #331