Fix decoder for EMU Prof II <!-- Thanks for submitting a pull request. Please fill the template below, otherwise we will not be able to process this pull request. -->
fixed decoding the serial number of a meter
...

<!-- What are the changes made in this pull request? -->

- fixed decoder
- added another test-case for the codec to validate the change

<!--
NOTE: This section is optional.

Motivate briefly why it is implemented this way, if that deviates from the
implementation proposal in the referenced issues.
- How should your reviewers approach this pull request?
- @mention reviewers with special requests or questions for them
-->

...

<!--
NOTE: This section is optional.

Any notes that we need to include in the Release Notes for the next release.
These notes are formatted as bullet points, written in past tense, and will be
combined with the labels of this Pull Request.

Always mention changes in API.
-->

- ...