Update Tektelic SRS decoder for new battery field
The latest version of Tektelic Comfort and Vivid sensors sends the battery level in the same way as other Tektelic sensors. This needs a change to the decoder as it now sends a 0x00 0xBA field instead of 0x00 0xFF.

I've modified the decoder to support this but note I haven't tested it as I can't paste in a decoder this long.